+++ CHANGES.rst
@@ -33,6 +33,8 @@ Release History
   `#41 <https://github.com/nengo/nengo-dl/issues/41>`_)
 - Fixed node outputs changing after simulator is built (fixes `#4
   <https://github.com/nengo/nengo-dl/issues/4>`__)
+- Better error message if user tries to train a network with non-differentiable
+  elements
 
 **Removed**